Refactored and functions sorted
